Update Readme.md
I found it quite confusing to have a header which said `FOR VUE 3` because my first thought was that this is for Vue 3, and then I noticed the little paragraph link below, which I think was well-hidden. I think this simple styling change makes it a bit easier on the eyes and the brain.